# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SwipeJobs API - Tinder/TikTok for Jobs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMyProfile
      tags: [users]
      summary: Get current user profile
      security:
        - clerkAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMyProfile
      tags: [users]
      summary: Update current user profile
      security:
        - clerkAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserProfileRequest"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/complete-onboarding:
    post:
      operationId: completeOnboarding
      tags: [users]
      summary: Complete user onboarding
      security:
        - clerkAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingRequest"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/push-token:
    post:
      operationId: registerPushToken
      tags: [users]
      summary: Register a push notification token
      security:
        - clerkAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushTokenRequest"
      responses:
        "200":
          description: Token registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /jobs:
    get:
      operationId: getJobs
      tags: [jobs]
      summary: Get job feed for swiping (auth optional — guests get all active jobs, authenticated users get personalized feed)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          description: Full-text search across job title, description, and tags
          schema:
            type: string
        - name: jobType
          in: query
          schema:
            type: string
            enum: [fulltime, contract, parttime, internship]
        - name: experienceLevel
          in: query
          schema:
            type: string
            enum: [junior, mid, senior, lead, executive, any]
        - name: remoteType
          in: query
          schema:
            type: string
            enum: [remote, hybrid, onsite]
      responses:
        "200":
          description: Job feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsResponse"
  /jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job details (auth optional)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
  /jobs/{jobId}/swipe:
    post:
      operationId: swipeJob
      tags: [jobs]
      summary: Swipe on a job (left=dismiss, right=save)
      security:
        - clerkAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwipeRequest"
      responses:
        "200":
          description: Swipe recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwipeResponse"
  /jobs/{jobId}/click:
    post:
      operationId: trackJobClick
      tags: [jobs]
      summary: Track an apply click on a job (auth optional)
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClickRequest"
      responses:
        "200":
          description: Click tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /employer/jobs:
    get:
      operationId: getEmployerJobs
      tags: [employer]
      summary: Get all jobs posted by this employer
      security:
        - clerkAuth: []
      responses:
        "200":
          description: Employer job listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerJobsResponse"
    post:
      operationId: createEmployerJob
      tags: [employer]
      summary: Create a new job posting (requires Stripe payment)
      security:
        - clerkAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Job created (pending payment)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployerJobCreatedResponse"
...
  /admin/jobs:
    get:
      operationId: adminGetJobs
      tags: [admin]
      summary: Admin - list all jobs
      security:
        - clerkAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Job list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobsResponse"
    post:
      operationId: adminCreateJob
      tags: [admin]
      summary: Admin - create a job
      security:
        - clerkAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobRequest"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /admin/jobs/{jobId}:
    patch:
      operationId: adminUpdateJob
      tags: [admin]
      summary: Admin - update a job
      security:
        - clerkAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequest"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: adminDeleteJob
      tags: [admin]
      summary: Admin - delete a job
      security:
        - clerkAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        name:
          type: string
          nullable: true
        email:
          type: string
        role:
          type: string
          enum: [job_seeker, employer, admin]
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive]
          nullable: true
        preferredLocation:
          type: string
          nullable: true
        remotePreference:
          type: string
          enum: [remote, hybrid, onsite, any]
          nullable: true
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        jobCategories:
          type: array
          items:
            type: string
        onboardingCompleted:
          type: boolean
        isAdmin:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - clerkId
        - email
        - role
        - onboardingCompleted
        - isAdmin
        - jobCategories
        - createdAt
    UpdateUserProfileRequest:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [job_seeker, employer]
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive]
        preferredLocation:
          type: string
        remotePreference:
          type: string
          enum: [remote, hybrid, onsite, any]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        jobCategories:
          type: array
          items:
            type: string
    OnboardingRequest:
      type: object
      properties:
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive]
        preferredLocation:
          type: string
        remotePreference:
          type: string
          enum: [remote, hybrid, onsite, any]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        jobCategories:
          type: array
          items:
            type: string
      required:
        - experienceLevel
        - remotePreference
        - jobCategories
    PushTokenRequest:
      type: object
      properties:
        token:
          type: string
        platform:
          type: string
          enum: [expo, apns, fcm]
      required:
        - token
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logoUrl:
          type: string
          nullable: true
        website:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required:
        - id
        - name
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          $ref: "#/components/schemas/Company"
        location:
          type: string
          nullable: true
        remoteType:
          type: string
          enum: [remote, hybrid, onsite]
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        jobType:
          type: string
          enum: [fulltime, parttime, contract, internship]
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive, any]
        tags:
          type: array
          items:
            type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        applyUrl:
          type: string
        source:
          type: string
        matchScore:
          type: integer
          nullable: true
          description: Match percentage 0-100 based on user preferences
        isSaved:
          type: boolean
        isPaidListing:
          type: boolean
        viewCount:
          type: integer
        createdAt:
          type: string
        expiresAt:
          type: string
          nullable: true
      required:
        - id
        - title
        - company
        - remoteType
        - jobType
        - experienceLevel
        - tags
        - shortDescription
        - fullDescription
        - applyUrl
        - source
        - isSaved
        - isPaidListing
        - viewCount
        - createdAt
    SwipeRequest:
      type: object
      properties:
        direction:
          type: string
          enum: [left, right]
      required:
        - direction
    SwipeResponse:
      type: object
      properties:
        success:
          type: boolean
        saved:
          type: boolean
      required:
        - success
        - saved
    ClickRequest:
      type: object
      properties:
        source:
          type: string
          enum: [apply_button, card_tap]
    JobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        hasMore:
          type: boolean
      required:
        - jobs
        - total
        - hasMore
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    EmployerJobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/EmployerJobListing"
        total:
          type: integer
      required:
        - jobs
        - total
    EmployerJobListing:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          type: string
          nullable: true
        remoteType:
          type: string
        jobType:
          type: string
        experienceLevel:
          type: string
        isActive:
          type: boolean
        isPaidListing:
          type: boolean
        viewCount:
          type: integer
        saveCount:
          type: integer
        clickCount:
          type: integer
        createdAt:
          type: string
        expiresAt:
          type: string
          nullable: true
      required:
        - id
        - title
        - remoteType
        - jobType
        - experienceLevel
        - isActive
        - isPaidListing
        - viewCount
        - saveCount
        - clickCount
        - createdAt
    EmployerJobCreatedResponse:
      type: object
      properties:
        jobId:
          type: integer
        checkoutUrl:
          type: string
          nullable: true
        message:
          type: string
      required:
        - jobId
        - message
...
    CreateJobRequest:
      type: object
      properties:
        title:
          type: string
        companyId:
          type: integer
        companyName:
          type: string
        location:
          type: string
        remoteType:
          type: string
          enum: [remote, hybrid, onsite]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        jobType:
          type: string
          enum: [fulltime, parttime, contract, internship]
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive, any]
        tags:
          type: array
          items:
            type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        applyUrl:
          type: string
      required:
        - title
        - remoteType
        - jobType
        - experienceLevel
        - shortDescription
        - fullDescription
        - applyUrl
    UpdateJobRequest:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        remoteType:
          type: string
          enum: [remote, hybrid, onsite]
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        jobType:
          type: string
          enum: [fulltime, parttime, contract, internship]
        experienceLevel:
          type: string
          enum: [junior, mid, senior, lead, executive, any]
        tags:
          type: array
          items:
            type: string
        shortDescription:
          type: string
        fullDescription:
          type: string
        applyUrl:
          type: string
        isActive:
          type: boolean
...