# lib/api-zod/src/generated/types/clickRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwipeJobs API - Tinder/TikTok for Jobs
 * OpenAPI spec version: 0.1.0
 */
import type { ClickRequestSource } from './clickRequestSource';

export interface ClickRequest {
  source?: ClickRequestSource;
}