# lib/api-zod/src/generated/types/updateUserProfileRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SwipeJobs API - Tinder/TikTok for Jobs
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateUserProfileRequestExperienceLevel } from './updateUserProfileRequestExperienceLevel';
import type { UpdateUserProfileRequestRemotePreference } from './updateUserProfileRequestRemotePreference';
import type { UpdateUserProfileRequestRole } from './updateUserProfileRequestRole';

export interface UpdateUserProfileRequest {
  name?: string;
  role?: UpdateUserProfileRequestRole;
  experienceLevel?: UpdateUserProfileRequestExperienceLevel;
  preferredLocation?: string;
  remotePreference?: UpdateUserProfileRequestRemotePreference;
  salaryMin?: number;
  salaryMax?: number;
  jobCategories?: string[];
}